+ MovieStats (Removed duplication but does allow for ones which slightly different names) + Added the ordering for MovieStats - Removed Testing Class
